+++ src/main/gen/io/runescript/plugin/lang/psi/RsScopedVariableExpression.java
@@ -4,8 +4,10 @@ package io.runescript.plugin.lang.psi;
 import java.util.List;
 import org.jetbrains.annotations.*;
 import com.intellij.psi.PsiElement;
+import com.intellij.psi.StubBasedPsiElement;
+import io.runescript.plugin.lang.stubs.RsScopedVariableExpressionStub;
 
-public interface RsScopedVariableExpression extends RsExpression {
+public interface RsScopedVariableExpression extends RsExpression, RsNamedElement, StubBasedPsiElement<RsScopedVariableExpressionStub> {
 
   @NotNull
   RsNameLiteral getNameLiteral();