+++ src/main/kotlin/io/runescript/plugin/ide/usages/RsReadWriteAccessDetector.kt
@@ -3,10 +3,9 @@ package io.runescript.plugin.ide.usages
 import com.intellij.codeInsight.highlighting.ReadWriteAccessDetector
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiReference
-import io.runescript.plugin.lang.psi.RsAssignmentStatement
-import io.runescript.plugin.lang.psi.RsDynamicExpression
-import io.runescript.plugin.lang.psi.RsLocalVariableDeclarationStatement
-import io.runescript.plugin.lang.psi.RsLocalVariableExpression
+import io.runescript.plugin.lang.psi.*
+import io.runescript.plugin.symbollang.psi.RsSymSymbol
+import io.runescript.plugin.symbollang.psi.isVarFile
 
 class RsReadWriteAccessDetector : ReadWriteAccessDetector() {
 
@@ -14,11 +13,14 @@ class RsReadWriteAccessDetector : ReadWriteAccessDetector() {
     //  without initializer are write accesses.
 
     override fun isReadWriteAccessible(element: PsiElement): Boolean {
+        if (element is RsSymSymbol) {
+            return element.containingFile?.virtualFile?.isVarFile() ?: false
+        }
         return element is RsLocalVariableExpression
     }
 
     override fun isDeclarationWriteAccess(element: PsiElement): Boolean {
-        require(element is RsLocalVariableExpression)
+        require(element is RsLocalVariableExpression || element is RsSymSymbol)
         val parent = element.parent
         return element is RsAssignmentStatement || parent is RsLocalVariableDeclarationStatement
     }
@@ -28,7 +30,9 @@ class RsReadWriteAccessDetector : ReadWriteAccessDetector() {
     }
 
     override fun getExpressionAccess(expression: PsiElement): Access {
-        require(expression is RsLocalVariableExpression || expression is RsDynamicExpression)
+        require(expression is RsLocalVariableExpression
+                || expression is RsScopedVariableExpression
+                || expression is RsDynamicExpression)
         val parent = expression.parent
         if (parent is RsAssignmentStatement || parent is RsLocalVariableDeclarationStatement) {
             return Access.Write