+++ src/main/kotlin/io/runescript/plugin/symbollang/psi/RsSymPsiExtensions.kt
@@ -3,3 +3,4 @@ package io.runescript.plugin.symbollang.psi
 import com.intellij.openapi.vfs.VirtualFile
 
 fun VirtualFile.isConstantFile() = name == "constant.sym"
+fun VirtualFile.isVarFile() = name == "varp.sym" || name == "varc.sym"